DRY AuthorityOrigin + AssetProcessor impls
We move them to runtime-common
