+++ runtime/common/Cargo.toml
@@ -61,6 +61,7 @@ std = [
     "pallet-anchors/std",
     "common-traits/std"
 ]
+runtime-benchmarks = []
 
 on-chain-release-build = [
     "sp-api/disable-logging",
