+++ runtime/development/src/lib.rs
@@ -65,7 +65,6 @@ use pallet_restricted_tokens::{
 pub use runtime_common::{Index, *};
 
 use chainbridge::constants::DEFAULT_RELAYER_VOTE_THRESHOLD;
-use frame_support::traits::{EnsureOrigin, EnsureOriginWithArg};
 
 pub mod xcm;
 pub use crate::xcm::*;
@@ -1218,66 +1217,10 @@ impl orml_asset_registry::Config for Runtime {
 	type Event = Event;
 	type CustomMetadata = CustomMetadata;
 	type AssetId = CurrencyId;
-	type AuthorityOrigin = AssetAuthority;
-	type AssetProcessor = CustomAssetProcessor;
+	type AuthorityOrigin = asset_registry::AuthorityOrigin<Origin>;
+	type AssetProcessor = asset_registry::CustomAssetProcessor;
 	type Balance = Balance;
 	type WeightInfo = ();
-}
-
-pub struct AssetAuthority;
-impl EnsureOriginWithArg<Origin, Option<CurrencyId>> for AssetAuthority {
-	type Success = ();
-
-	fn try_origin(origin: Origin, asset_id: &Option<CurrencyId>) -> Result<Self::Success, Origin> {
-		match asset_id {
-			// Only the pools pallet should directly register/update tranche tokens
-			Some(CurrencyId::Tranche(_, _)) => Err(origin),
-
-			// Any other `asset_id` defaults to EnsureRoot
-			_ => EnsureRoot::try_origin(origin),
-		}
-	}
-
-	#[cfg(feature = "runtime-benchmarks")]
-	fn successful_origin(_: &Option<CurrencyId>) -> Origin {
-		todo!()
-	}
-}
-
-#[derive(
-	Clone, Copy, PartialOrd, Ord, PartialEq, Eq, Debug, Encode, Decode, TypeInfo, MaxEncodedLen,
-)]
-pub struct CustomAssetProcessor {}
-
-use orml_traits::asset_registry::AssetProcessor;
-
-impl AssetProcessor<CurrencyId, orml_asset_registry::AssetMetadata<Balance, CustomMetadata>>
-	for CustomAssetProcessor
-{
-	fn pre_register(
-		id: Option<CurrencyId>,
-		metadata: orml_asset_registry::AssetMetadata<Balance, CustomMetadata>,
-	) -> Result<
-		(
-			CurrencyId,
-			orml_asset_registry::AssetMetadata<Balance, CustomMetadata>,
-		),
-		sp_runtime::DispatchError,
-	> {
-		match id {
-			Some(id) => Ok((id, metadata)),
-			None => Err(sp_runtime::DispatchError::Other(
-				"asset-registry: AssetId is required",
-			)),
-		}
-	}
-
-	fn post_register(
-		_id: CurrencyId,
-		_asset_metadata: orml_asset_registry::AssetMetadata<Balance, CustomMetadata>,
-	) -> Result<(), sp_runtime::DispatchError> {
-		Ok(())
-	}
 }
 
 parameter_types! {
